Rename script with a more general name and improve the trailets
