+++ lstchain/scripts/onsite/onsite_fit_intensity_scan.py
@@ -2,7 +2,7 @@
 
 """
 
- Onsite script to fit a filter scan production
+ Onsite script to fit a intensity scan production
 
 
 """
@@ -24,7 +24,7 @@ parser = argparse.ArgumentParser(description='Create flat-field calibration file
 required = parser.add_argument_group('required arguments')
 optional = parser.add_argument_group('optional arguments')
 
-required.add_argument('-d', '--date', help="Date of the filter scan (YYYYMMDD)", required=True)
+required.add_argument('-d', '--date', help="Date of the scan (YYYYMMDD)", required=True)
 
 # config file is mandatory because it contains the list of input runs
 required.add_argument('-c','--config', help="Config file (json format) with the list of runs", required=True)
@@ -34,7 +34,6 @@ optional.add_argument('-v', '--prod_version', help="Version of the production",
                       default=f"v{version}")
 optional.add_argument('-b','--base_dir', help="Root dir for the output directory tree", type=str, default='/fefs/aswg/data/real')
 optional.add_argument('--sub_run', help="sub-run to be processed.", type=int, default=0)
-optional.add_argument('--gains', help="List of gains to consider (format: --gains 0 1)", type=int, nargs="+", default=[0,1])
 
 args = parser.parse_args()
 date = args.date
@@ -42,8 +41,6 @@ prod_id = args.prod_version
 base_dir = args.base_dir
 sub_run = args.sub_run
 config_file = args.config
-gains = args.gains
-
 
 def main():
 
@@ -64,7 +61,7 @@ def main():
         print(f"\n--> Input directory {input_dir}")
 
         # verify output dir
-        output_dir = f"{base_dir}/monitoring/PixelCalibration/filter_scan/{date}/{prod_id}"
+        output_dir = f"{base_dir}/monitoring/PixelCalibration/intensity_scan/{date}/{prod_id}"
         if not os.path.exists(output_dir):
             print(f"--> Create directory {output_dir}")
             os.makedirs(output_dir, exist_ok=True)
@@ -75,38 +72,33 @@ def main():
             print(f"--> Create directory {log_dir}")
             os.makedirs(log_dir, exist_ok=True)
 
-        for gain in gains:
+        # define charge file names
+        output_file = f"{output_dir}/intensity_scan_fit_{date}.{sub_run:04d}.h5"
+        log_file = f"{output_dir}/log/intensity_scan_fit_{date}.{sub_run:04d}.log"
 
-            print(f"\n-->>>>> Process {channel[gain]} gain <<<<<")
-            # define charge file names
-            output_file = f"{output_dir}/filter_scan_fit_{channel[gain]}_{date}.{sub_run:04d}.h5"
-            log_file = f"{output_dir}/log/filter_scan_fit_{channel[gain]}_{date}.{sub_run:04d}.log"
+        print(f"\n--> Output file {output_file}")
 
-            print(f"\n--> Output file {output_file}")
+        if os.path.exists(output_file):
+            if query_yes_no(">>> Output file exists already. Do you want to remove it?"):
+                os.remove(output_file)
+            else:
+                print(f"\n--> Stop")
+                exit(1)
 
-            if os.path.exists(output_file):
-                if query_yes_no(">>> Output file exists already. Do you want to remove it?"):
-                    os.remove(output_file)
-                else:
-                    print(f"\n--> Stop")
-                    exit(1)
+        print(f"\n--> Log file {log_file}")
 
-            print(f"\n--> Log file {log_file}")
+        #
+        # produce intensity scan fit file
+        #
 
-            #
-            # produce filter scan fit file
-            #
+        cmd = f"lstchain_fit_intensity_scan " \
+              f"--config={config_file} --input_dir={input_dir} --output_path={output_file} "\
+              f"--sub_run={sub_run} "\
+              f" >  {log_file} 2>&1"
 
-            cmd = f"lstchain_fit_filter_scan " \
-                  f"--config={config_file} --input_dir={input_dir} --output_path={output_file} "\
-                  f"--sub_run={sub_run} --gain_channel={gain} "\
-                  f" >  {log_file} 2>&1"
-
-            print("\n--> RUNNING...")
-            if os.system(cmd):
-                raise Exception(f"Error in command execution: {cmd}")
-
-
+        print("\n--> RUNNING...")
+        if os.system(cmd):
+            raise Exception(f"Error in command execution: {cmd}")
 
         print("\n--> END")
